Moved JUnit 4 integration classes from internal package to parent.
